Updates to Upload Service
Upload service now has two separate enums for both whether or not it is currently uploading as well as the status of the most recent upload. Added checks for whether or not wifi is connected as well. Various miscellaneous cleanup also happened. Still need to do some UI cleanup and notification cleanup but that can wait until Thirsday.